Add RtlCreateUserThread injection

+++ al-khaser/Code Injections/SetWindowsHooksEx.cpp
@@ -40,7 +40,7 @@ BOOL SetWindowsHooksEx_Injection()
 		print_last_error(_T("GetProcAddress"));
 		return FALSE;
 	}
-	_tprintf(_T("\t[+] Found at 0x%08x\n"), myFunctionAddress);
+	_tprintf(_T("\t[+] Found at 0x%08x\n"), (UINT)myFunctionAddress);
 
 	/* Injection happens here */
 	hHook = SetWindowsHookEx(WH_KEYBOARD, myFunctionAddress, hOurDll, dwThreadId);
